+++ src/Game/C/sound.c
@@ -505,7 +505,7 @@ void ComputeDoppler(CHANNEL_DATA *ch)
 	int seperationrate;
 
 	VECTOR *srcPos;
-	long* srcVel; // LONGVECTOR3
+	int* srcVel; // LONGVECTOR3
 
 	PLAYER *pl;
 	int dx, dy, dz;
@@ -518,7 +518,7 @@ void ComputeDoppler(CHANNEL_DATA *ch)
 		return;
 	}
 
-	srcVel = (long*)ch->srcvelocity;
+	srcVel = (int*)ch->srcvelocity;
 
 	pl = &player[ch->player];
 
@@ -1102,7 +1102,7 @@ int GetFreeChannel(int force)
 }
 
 // [D] [T]
-void AllocateReverb(long mode, long depth)
+void AllocateReverb(int mode, int depth)
 {
 	SpuReverbAttr r_attr;
